登录接口流程修改+APP端数据互通

+++ live-service/src/main/java/com/xq/live/service/SmsSendService.java
@@ -46,7 +46,7 @@ public interface SmsSendService {
      * @param inVo
      * @return
      */
-    Integer isVerify(SmsSendInVo inVo);
+    Long isVerify(SmsSendInVo inVo);
 
     /**
      * 商家端app注册验证码是否通过
